archive_20251116_audit/ml_backup_20251114_120646/requirements.txt
# ML Service Dependencies

# Core ML libraries
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.3.0
torch==2.0.1
transformers==4.30.2
onnx==1.14.0
onnxruntime==1.15.1
joblib==1.2.0

# MLOps
mlflow==2.4.1
prometheus-client==0.17.1

# API and Web
fastapi==0.98.0
uvicorn==0.22.0
pydantic==1.10.9

# Database and storage
sqlalchemy==2.0.12
psycopg2-binary==2.9.6
redis==4.5.5
influxdb-client==1.36.1

# Utilities
python-dotenv==1.0.0
celery==5.3.1
tqdm==4.65.0
loguru==0.7.0

# Testing
pytest==7.3.1
httpx==0.24.1